feat: dispatch compose
