+++ lib/api/api-pipeline.js
@@ -12,6 +12,7 @@ const {
 } = require('../core/errors')
 const util = require('../core/util')
 const { AsyncResource } = require('node:async_hooks')
+const redirect = require('../interceptor/redirect')
 const { addSignal, removeSignal } = require('./abort-signal')
 const assert = require('node:assert')
 
@@ -239,7 +240,11 @@ class PipelineHandler extends AsyncResource {
 function pipeline (opts, handler) {
   try {
     const pipelineHandler = new PipelineHandler(opts, handler)
-    this.dispatch({ ...opts, body: pipelineHandler.req }, pipelineHandler)
+
+    this
+      .compose(redirect(opts))
+      .dispatch({ ...opts, body: pipelineHandler.req }, pipelineHandler)
+
     return pipelineHandler.ret
   } catch (err) {
     return new PassThrough().destroy(err)
