+++ lib/api/api-stream.js
@@ -7,6 +7,7 @@ const {
   RequestAbortedError
 } = require('../core/errors')
 const util = require('../core/util')
+const redirect = require('../interceptor/redirect')
 const { getResolveErrorBodyCallback } = require('./util')
 const { AsyncResource } = require('node:async_hooks')
 const { addSignal, removeSignal } = require('./abort-signal')
@@ -207,7 +208,9 @@ function stream (opts, factory, callback) {
   }
 
   try {
-    this.dispatch(opts, new StreamHandler(opts, factory, callback))
+    this
+      .compose(redirect(opts))
+      .dispatch(opts, new StreamHandler(opts, factory, callback))
   } catch (err) {
     if (typeof callback !== 'function') {
       throw err
