+++ lib/api/api-upgrade.js
@@ -3,6 +3,7 @@
 const { InvalidArgumentError, RequestAbortedError, SocketError } = require('../core/errors')
 const { AsyncResource } = require('node:async_hooks')
 const util = require('../core/util')
+const redirect = require('../interceptor/redirect')
 const { addSignal, removeSignal } = require('./abort-signal')
 const assert = require('node:assert')
 
@@ -87,12 +88,9 @@ function upgrade (opts, callback) {
   }
 
   try {
-    const upgradeHandler = new UpgradeHandler(opts, callback)
-    this.dispatch({
-      ...opts,
-      method: opts.method || 'GET',
-      upgrade: opts.protocol || 'Websocket'
-    }, upgradeHandler)
+    this
+      .compose(redirect(opts))
+      .dispatch({ ...opts, method: opts?.method || 'GET', upgrade: opts?.protocol || 'Websocket' }, new UpgradeHandler(opts, callback))
   } catch (err) {
     if (typeof callback !== 'function') {
       throw err
