+++ lib/dispatcher/client.js
@@ -68,7 +68,6 @@ const {
   kClose,
   kDestroy,
   kDispatch,
-  kInterceptors,
   kLocalAddress,
   kMaxResponseSize,
   kHTTPConnVersion,
@@ -136,7 +135,6 @@ class Client extends DispatcherBase {
    * @param {import('../../types/client.js').Client.Options} options
    */
   constructor (url, {
-    interceptors,
     maxHeaderSize,
     headersTimeout,
     socketTimeout,
@@ -223,10 +221,6 @@ class Client extends DispatcherBase {
       throw new InvalidArgumentError('connect must be a function or an object')
     }
 
-    if (maxRedirections != null && (!Number.isInteger(maxRedirections) || maxRedirections < 0)) {
-      throw new InvalidArgumentError('maxRedirections must be a positive number')
-    }
-
     if (maxRequestsPerClient != null && (!Number.isInteger(maxRequestsPerClient) || maxRequestsPerClient < 0)) {
       throw new InvalidArgumentError('maxRequestsPerClient must be a positive number')
     }
@@ -267,9 +261,6 @@ class Client extends DispatcherBase {
       })
     }
 
-    this[kInterceptors] = interceptors?.Client && Array.isArray(interceptors.Client)
-      ? interceptors.Client
-      : [createRedirectInterceptor({ maxRedirections })]
     this[kUrl] = util.parseOrigin(url)
     this[kConnector] = connect
     this[kSocket] = null
@@ -488,7 +479,6 @@ function onHTTP2GoAway (code) {
 }
 
 const constants = require('../llhttp/constants.js')
-const createRedirectInterceptor = require('../interceptor/redirectInterceptor.js')
 const EMPTY_BUF = Buffer.alloc(0)
 
 async function lazyllhttp () {
