+++ lib/dispatcher/dispatcher.js
@@ -2,7 +2,11 @@
 
 const EventEmitter = require('node:events')
 
+const kDispatcherVersion = Symbol.for('undici.dispatcher.version')
+
 class Dispatcher extends EventEmitter {
+  [kDispatcherVersion] = 1
+
   dispatch () {
     throw new Error('not implemented')
   }
@@ -13,6 +17,26 @@ class Dispatcher extends EventEmitter {
 
   destroy () {
     throw new Error('not implemented')
+  }
+
+  compose (...interceptors) {
+    let dispatcher = this
+    for (const interceptor of interceptors) {
+      if (interceptor == null) {
+        continue
+      }
+
+      if (typeof interceptor !== 'function') {
+        throw new Error('invalid interceptor')
+      }
+
+      dispatcher = interceptor(dispatcher) ?? dispatcher
+
+      if (dispatcher[kDispatcherVersion] !== 1) {
+        throw new Error('invalid dispatcher')
+      }
+    }
+    return dispatcher
   }
 }
