+++ lib/dispatcher/pool.js
@@ -12,7 +12,7 @@ const {
   InvalidArgumentError
 } = require('../core/errors')
 const util = require('../core/util')
-const { kUrl, kInterceptors } = require('../core/symbols')
+const { kUrl } = require('../core/symbols')
 const buildConnector = require('../core/connect')
 
 const kOptions = Symbol('options')
@@ -63,15 +63,9 @@ class Pool extends PoolBase {
       })
     }
 
-    this[kInterceptors] = options.interceptors?.Pool && Array.isArray(options.interceptors.Pool)
-      ? options.interceptors.Pool
-      : []
     this[kConnections] = connections || null
     this[kUrl] = util.parseOrigin(origin)
     this[kOptions] = { ...util.deepClone(options), connect, allowH2 }
-    this[kOptions].interceptors = options.interceptors
-      ? { ...options.interceptors }
-      : undefined
     this[kFactory] = factory
   }
