+++ lib/dispatcher/proxy-agent.js
@@ -1,10 +1,10 @@
 'use strict'
 
-const { kProxy, kClose, kDestroy, kInterceptors } = require('../core/symbols')
+const { kProxy, kClose, kDestroy } = require('../core/symbols')
 const { URL } = require('node:url')
-const Agent = require('./agent')
-const Pool = require('./pool')
-const DispatcherBase = require('./dispatcher-base')
+const Agent = require('../agent')
+const Pool = require('../pool')
+const DispatcherBase = require('../dispatcher-base')
 const { InvalidArgumentError, RequestAbortedError } = require('../core/errors')
 const buildConnector = require('../core/connect')
 
@@ -41,9 +41,6 @@ class ProxyAgent extends DispatcherBase {
 
     this[kProxy] = { uri: href, protocol }
     this[kAgent] = new Agent(opts)
-    this[kInterceptors] = opts.interceptors?.ProxyAgent && Array.isArray(opts.interceptors.ProxyAgent)
-      ? opts.interceptors.ProxyAgent
-      : []
     this[kRequestTls] = opts.requestTls
     this[kProxyTls] = opts.proxyTls
     this[kProxyHeaders] = opts.headers || {}
@@ -180,4 +177,24 @@ function throwIfProxyAuthIsSent (headers) {
   }
 }
 
-module.exports = ProxyAgent
+module.exports = (opts) => {
+  if (typeof opts === 'string') {
+    opts = { uri: opts }
+  }
+
+  if (!opts || !opts.uri) {
+    throw new InvalidArgumentError('Proxy opts.uri is mandatory')
+  }
+
+  const { clientFactory = defaultFactory } = opts
+
+  if (typeof clientFactory !== 'function') {
+    throw new InvalidArgumentError('Proxy opts.clientFactory must be a function.')
+  }
+
+  if (opts.auth && opts.token) {
+    throw new InvalidArgumentError('opts.auth cannot be used in combination with opts.token')
+  }
+
+  return (dispatcher) => new ProxyAgent(dispatcher, opts)
+}
