+++ test/mock-pool.js
@@ -123,16 +123,6 @@ test('MockPool - intercept should return a MockInterceptor', (t) => {
 })
 
 describe('MockPool - intercept validation', () => {
-  test('it should error if no options specified in the intercept', t => {
-    t = tspl(t, { plan: 1 })
-    const mockAgent = new MockAgent()
-    after(() => mockAgent.close())
-
-    const mockPool = mockAgent.get('http://localhost:9999')
-
-    t.throws(() => mockPool.intercept(), new InvalidArgumentError('opts must be an object'))
-  })
-
   test('it should error if no path specified in the intercept', t => {
     t = tspl(t, { plan: 1 })
     const mockAgent = new MockAgent()
