+++ test/retry-handler.js
@@ -5,7 +5,8 @@ const { test, after } = require('node:test')
 const { createServer } = require('node:http')
 const { once } = require('node:events')
 
-const { RetryHandler, Client } = require('..')
+const { Client } = require('..')
+const retry = require('../lib/interceptor/retry')
 const { RequestHandler } = require('../lib/api/api-request')
 
 test('Should retry status code', async t => {
@@ -14,28 +15,6 @@ test('Should retry status code', async t => {
   let counter = 0
   const chunks = []
   const server = createServer()
-  const dispatchOptions = {
-    retryOptions: {
-      retry: (err, { state, opts }, done) => {
-        counter++
-
-        if (
-          err.statusCode === 500 ||
-          err.message.includes('other side closed')
-        ) {
-          setTimeout(done, 500)
-          return
-        }
-
-        return done(err)
-      }
-    },
-    method: 'GET',
-    path: '/',
-    headers: {
-      'content-type': 'application/json'
-    }
-  }
 
   server.on('request', (req, res) => {
     switch (counter) {
@@ -57,29 +36,26 @@ test('Should retry status code', async t => {
 
   server.listen(0, () => {
     const client = new Client(`http://localhost:${server.address().port}`)
-    const handler = new RetryHandler(dispatchOptions, {
-      dispatch: client.dispatch.bind(client),
-      handler: {
-        onConnect () {
-          t.ok(true, 'pass')
-        },
-        onHeaders (status, _rawHeaders, resume, _statusMessage) {
-          t.strictEqual(status, 200)
-          return true
-        },
-        onData (chunk) {
-          chunks.push(chunk)
-          return true
-        },
-        onComplete () {
-          t.strictEqual(Buffer.concat(chunks).toString('utf-8'), 'hello world!')
-          t.strictEqual(counter, 2)
-        },
-        onError () {
-          t.fail()
-        }
+    const handler = {
+      onConnect () {
+        t.ok(true, 'pass')
+      },
+      onHeaders (status, _rawHeaders, resume, _statusMessage) {
+        t.strictEqual(status, 200)
+        return true
+      },
+      onData (chunk) {
+        chunks.push(chunk)
+        return true
+      },
+      onComplete () {
+        t.strictEqual(Buffer.concat(chunks).toString('utf-8'), 'hello world!')
+        t.strictEqual(counter, 2)
+      },
+      onError () {
+        t.fail()
       }
-    })
+    }
 
     after(async () => {
       await client.close()
@@ -88,7 +64,21 @@ test('Should retry status code', async t => {
       await once(server, 'close')
     })
 
-    client.dispatch(
+    client.compose(retry({
+      retry: (err, { state, opts }, done) => {
+        counter++
+
+        if (
+          err.statusCode === 500 ||
+          err.message.includes('other side closed')
+        ) {
+          setTimeout(done, 500)
+          return
+        }
+
+        return done(err)
+      }
+    })).dispatch(
       {
         method: 'GET',
         path: '/',
@@ -110,13 +100,6 @@ test('Should use retry-after header for retries', async t => {
   const chunks = []
   const server = createServer()
   let checkpoint
-  const dispatchOptions = {
-    method: 'PUT',
-    path: '/',
-    headers: {
-      'content-type': 'application/json'
-    }
-  }
 
   server.on('request', (req, res) => {
     switch (counter) {
@@ -141,28 +124,25 @@ test('Should use retry-after header for retries', async t => {
 
   server.listen(0, () => {
     const client = new Client(`http://localhost:${server.address().port}`)
-    const handler = new RetryHandler(dispatchOptions, {
-      dispatch: client.dispatch.bind(client),
-      handler: {
-        onConnect () {
-          t.ok(true, 'pass')
-        },
-        onHeaders (status, _rawHeaders, resume, _statusMessage) {
-          t.strictEqual(status, 200)
-          return true
-        },
-        onData (chunk) {
-          chunks.push(chunk)
-          return true
-        },
-        onComplete () {
-          t.strictEqual(Buffer.concat(chunks).toString('utf-8'), 'hello world!')
-        },
-        onError (err) {
-          t.ifError(err)
-        }
+    const handler = {
+      onConnect () {
+        t.ok(true, 'pass')
+      },
+      onHeaders (status, _rawHeaders, resume, _statusMessage) {
+        t.strictEqual(status, 200)
+        return true
+      },
+      onData (chunk) {
+        chunks.push(chunk)
+        return true
+      },
+      onComplete () {
+        t.strictEqual(Buffer.concat(chunks).toString('utf-8'), 'hello world!')
+      },
+      onError (err) {
+        t.ifError(err)
       }
-    })
+    }
 
     after(async () => {
       await client.close()
@@ -171,7 +151,9 @@ test('Should use retry-after header for retries', async t => {
       await once(server, 'close')
     })
 
-    client.dispatch(
+    client.compose(retry({
+
+    })).dispatch(
       {
         method: 'PUT',
         path: '/',
@@ -193,13 +175,6 @@ test('Should use retry-after header for retries (date)', async t => {
   const chunks = []
   const server = createServer()
   let checkpoint
-  const dispatchOptions = {
-    method: 'PUT',
-    path: '/',
-    headers: {
-      'content-type': 'application/json'
-    }
-  }
 
   server.on('request', (req, res) => {
     switch (counter) {
@@ -226,28 +201,25 @@ test('Should use retry-after header for retries (date)', async t => {
 
   server.listen(0, () => {
     const client = new Client(`http://localhost:${server.address().port}`)
-    const handler = new RetryHandler(dispatchOptions, {
-      dispatch: client.dispatch.bind(client),
-      handler: {
-        onConnect () {
-          t.ok(true, 'pass')
-        },
-        onHeaders (status, _rawHeaders, resume, _statusMessage) {
-          t.strictEqual(status, 200)
-          return true
-        },
-        onData (chunk) {
-          chunks.push(chunk)
-          return true
-        },
-        onComplete () {
-          t.strictEqual(Buffer.concat(chunks).toString('utf-8'), 'hello world!')
-        },
-        onError (err) {
-          t.ifError(err)
-        }
+    const handler = {
+      onConnect () {
+        t.ok(true, 'pass')
+      },
+      onHeaders (status, _rawHeaders, resume, _statusMessage) {
+        t.strictEqual(status, 200)
+        return true
+      },
+      onData (chunk) {
+        chunks.push(chunk)
+        return true
+      },
+      onComplete () {
+        t.strictEqual(Buffer.concat(chunks).toString('utf-8'), 'hello world!')
+      },
+      onError (err) {
+        t.ifError(err)
       }
-    })
+    }
 
     after(async () => {
       await client.close()
@@ -256,7 +228,8 @@ test('Should use retry-after header for retries (date)', async t => {
       await once(server, 'close')
     })
 
-    client.dispatch(
+    client.compose(retry({
+    })).dispatch(
       {
         method: 'PUT',
         path: '/',
@@ -277,13 +250,6 @@ test('Should retry with defaults', async t => {
   let counter = 0
   const chunks = []
   const server = createServer()
-  const dispatchOptions = {
-    method: 'GET',
-    path: '/',
-    headers: {
-      'content-type': 'application/json'
-    }
-  }
 
   server.on('request', (req, res) => {
     switch (counter) {
@@ -308,28 +274,25 @@ test('Should retry with defaults', async t => {
 
   server.listen(0, () => {
     const client = new Client(`http://localhost:${server.address().port}`)
-    const handler = new RetryHandler(dispatchOptions, {
-      dispatch: client.dispatch.bind(client),
-      handler: {
-        onConnect () {
-          t.ok(true, 'pass')
-        },
-        onHeaders (status, _rawHeaders, resume, _statusMessage) {
-          t.strictEqual(status, 200)
-          return true
-        },
-        onData (chunk) {
-          chunks.push(chunk)
-          return true
-        },
-        onComplete () {
-          t.strictEqual(Buffer.concat(chunks).toString('utf-8'), 'hello world!')
-        },
-        onError (err) {
-          t.ifError(err)
-        }
+    const handler = {
+      onConnect () {
+        t.ok(true, 'pass')
+      },
+      onHeaders (status, _rawHeaders, resume, _statusMessage) {
+        t.strictEqual(status, 200)
+        return true
+      },
+      onData (chunk) {
+        chunks.push(chunk)
+        return true
+      },
+      onComplete () {
+        t.strictEqual(Buffer.concat(chunks).toString('utf-8'), 'hello world!')
+      },
+      onError (err) {
+        t.ifError(err)
       }
-    })
+    }
 
     after(async () => {
       await client.close()
@@ -338,7 +301,7 @@ test('Should retry with defaults', async t => {
       await once(server, 'close')
     })
 
-    client.dispatch(
+    client.compose(retry()).dispatch(
       {
         method: 'GET',
         path: '/',
@@ -379,8 +342,30 @@ test('Should handle 206 partial content', async t => {
     x++
   })
 
-  const dispatchOptions = {
-    retryOptions: {
+  server.listen(0, () => {
+    const client = new Client(`http://localhost:${server.address().port}`)
+    const handler = {
+      onConnect () {
+        t.ok(true, 'pass')
+      },
+      onHeaders (status, _rawHeaders, resume, _statusMessage) {
+        t.strictEqual(status, 200)
+        return true
+      },
+      onData (chunk) {
+        chunks.push(chunk)
+        return true
+      },
+      onComplete () {
+        t.strictEqual(Buffer.concat(chunks).toString('utf-8'), 'abcdef')
+        t.strictEqual(counter, 1)
+      },
+      onError () {
+        t.fail()
+      }
+    }
+
+    client.compose(retry({
       retry: function (err, _, done) {
         counter++
 
@@ -392,43 +377,7 @@ test('Should handle 206 partial content', async t => {
 
         setTimeout(done, 800)
       }
-    },
-    method: 'GET',
-    path: '/',
-    headers: {
-      'content-type': 'application/json'
-    }
-  }
-
-  server.listen(0, () => {
-    const client = new Client(`http://localhost:${server.address().port}`)
-    const handler = new RetryHandler(dispatchOptions, {
-      dispatch: (...args) => {
-        return client.dispatch(...args)
-      },
-      handler: {
-        onConnect () {
-          t.ok(true, 'pass')
-        },
-        onHeaders (status, _rawHeaders, resume, _statusMessage) {
-          t.strictEqual(status, 200)
-          return true
-        },
-        onData (chunk) {
-          chunks.push(chunk)
-          return true
-        },
-        onComplete () {
-          t.strictEqual(Buffer.concat(chunks).toString('utf-8'), 'abcdef')
-          t.strictEqual(counter, 1)
-        },
-        onError () {
-          t.fail()
-        }
-      }
-    })
-
-    client.dispatch(
+    })).dispatch(
       {
         method: 'GET',
         path: '/',
@@ -475,46 +424,30 @@ test('Should handle 206 partial content - bad-etag', async t => {
     x++
   })
 
-  const dispatchOptions = {
-    method: 'GET',
-    path: '/',
-    headers: {
-      'content-type': 'application/json'
-    }
-  }
-
   server.listen(0, () => {
     const client = new Client(`http://localhost:${server.address().port}`)
-    const handler = new RetryHandler(
-      dispatchOptions,
-      {
-        dispatch: (...args) => {
-          return client.dispatch(...args)
-        },
-        handler: {
-          onConnect () {
-            t.ok(true, 'pass')
-          },
-          onHeaders (status, _rawHeaders, resume, _statusMessage) {
-            t.ok(true, 'pass')
-            return true
-          },
-          onData (chunk) {
-            chunks.push(chunk)
-            return true
-          },
-          onComplete () {
-            t.ifError('should not complete')
-          },
-          onError (err) {
-            t.strictEqual(Buffer.concat(chunks).toString('utf-8'), 'abc')
-            t.strictEqual(err.code, 'UND_ERR_REQ_RETRY')
-          }
-        }
+    const handler = {
+      onConnect () {
+        t.ok(true, 'pass')
+      },
+      onHeaders (status, _rawHeaders, resume, _statusMessage) {
+        t.ok(true, 'pass')
+        return true
+      },
+      onData (chunk) {
+        chunks.push(chunk)
+        return true
+      },
+      onComplete () {
+        t.ifError('should not complete')
+      },
+      onError (err) {
+        t.strictEqual(Buffer.concat(chunks).toString('utf-8'), 'abc')
+        t.strictEqual(err.code, 'UND_ERR_REQ_RETRY')
       }
-    )
+    }
 
-    client.dispatch(
+    client.compose(retry()).dispatch(
       {
         method: 'GET',
         path: '/',
@@ -539,29 +472,6 @@ test('Should handle 206 partial content - bad-etag', async t => {
 test('retrying a request with a body', async t => {
   let counter = 0
   const server = createServer()
-  const dispatchOptions = {
-    retryOptions: {
-      retry: (err, { state, opts }, done) => {
-        counter++
-
-        if (
-          err.statusCode === 500 ||
-          err.message.includes('other side closed')
-        ) {
-          setTimeout(done, 500)
-          return
-        }
-
-        return done(err)
-      }
-    },
-    method: 'POST',
-    path: '/',
-    headers: {
-      'content-type': 'application/json'
-    },
-    body: JSON.stringify({ hello: 'world' })
-  }
 
   t = tspl(t, { plan: 1 })
 
@@ -585,11 +495,15 @@ test('retrying a request with a body', async t => {
 
   server.listen(0, () => {
     const client = new Client(`http://localhost:${server.address().port}`)
-    const handler = new RetryHandler(dispatchOptions, {
-      dispatch: client.dispatch.bind(client),
-      handler: new RequestHandler(dispatchOptions, (err, data) => {
-        t.ifError(err)
-      })
+    const handler = new RequestHandler({
+      method: 'POST',
+      path: '/',
+      headers: {
+        'content-type': 'application/json'
+      },
+      body: JSON.stringify({ hello: 'world' })
+    }, (err, data) => {
+      t.ifError(err)
     })
 
     after(async () => {
@@ -599,7 +513,21 @@ test('retrying a request with a body', async t => {
       await once(server, 'close')
     })
 
-    client.dispatch(
+    client.compose(retry({
+      retry: (err, { state, opts }, done) => {
+        counter++
+
+        if (
+          err.statusCode === 500 ||
+          err.message.includes('other side closed')
+        ) {
+          setTimeout(done, 500)
+          return
+        }
+
+        return done(err)
+      }
+    })).dispatch(
       {
         method: 'POST',
         path: '/',
@@ -624,42 +552,29 @@ test('should not error if request is not meant to be retried', async t => {
     res.end('Bad request')
   })
 
-  const dispatchOptions = {
-    retryOptions: {
-      method: 'GET',
-      path: '/',
-      headers: {
-        'content-type': 'application/json'
-      }
-    }
-  }
-
   server.listen(0, () => {
     const client = new Client(`http://localhost:${server.address().port}`)
     const chunks = []
-    const handler = new RetryHandler(dispatchOptions, {
-      dispatch: client.dispatch.bind(client),
-      handler: {
-        onConnect () {
-          t.ok(true, 'pass')
-        },
-        onHeaders (status, _rawHeaders, resume, _statusMessage) {
-          t.strictEqual(status, 400)
-          return true
-        },
-        onData (chunk) {
-          chunks.push(chunk)
-          return true
-        },
-        onComplete () {
-          t.strictEqual(Buffer.concat(chunks).toString('utf-8'), 'Bad request')
-        },
-        onError (err) {
-          console.log({ err })
-          t.fail()
-        }
+    const handler = {
+      onConnect () {
+        t.ok(true, 'pass')
+      },
+      onHeaders (status, _rawHeaders, resume, _statusMessage) {
+        t.strictEqual(status, 400)
+        return true
+      },
+      onData (chunk) {
+        chunks.push(chunk)
+        return true
+      },
+      onComplete () {
+        t.strictEqual(Buffer.concat(chunks).toString('utf-8'), 'Bad request')
+      },
+      onError (err) {
+        console.log({ err })
+        t.fail()
       }
-    })
+    }
 
     after(async () => {
       await client.close()
@@ -668,7 +583,7 @@ test('should not error if request is not meant to be retried', async t => {
       await once(server, 'close')
     })
 
-    client.dispatch(
+    client.compose(retry()).dispatch(
       {
         method: 'GET',
         path: '/',
