+++ test/types/client.test-d.ts
@@ -51,9 +51,6 @@ expectAssignable<Client>(new Client(new URL('http://localhost'), {}))
     maxCachedSessions: 1
   }))
   expectAssignable<Client>(new Client('', {
-    maxRedirections: 1
-  }))
-  expectAssignable<Client>(new Client('', {
     maxRequestsPerClient: 1
   }))
   expectAssignable<Client>(new Client('', {
@@ -67,18 +64,6 @@ expectAssignable<Client>(new Client(new URL('http://localhost'), {}))
   }))
   expectAssignable<Client>(new Client('', {
     autoSelectFamilyAttemptTimeout: 300e3
-  }))
-  expectAssignable<Client>(new Client('', {
-    interceptors: {
-      Client: [(dispatcher) => {
-        expectAssignable<Dispatcher['dispatch']>(dispatcher);
-        return (opts, handlers) => {
-          expectAssignable<Dispatcher.DispatchOptions>(opts);
-          expectAssignable<Dispatcher.DispatchHandlers>(handlers);
-          return dispatcher(opts, handlers)
-        }
-      }]
-    }
   }))
 }
