+++ types/client.d.ts
@@ -32,8 +32,6 @@ export declare namespace Client {
     Client: readonly Dispatcher.DispatchInterceptor[];
   }
   export interface Options {
-    /** TODO */
-    interceptors?: OptionsInterceptors;
     /** The maximum length of request headers in bytes. Default: Node.js' `--max-http-header-size` or `16384` (16KiB). */
     maxHeaderSize?: number;
     /** The amount of time, in milliseconds, the parser will wait to receive the complete HTTP headers (Node 14 and above only). Default: `300e3` milliseconds (300s). */
@@ -68,8 +66,6 @@ export declare namespace Client {
     strictContentLength?: boolean;
     /** TODO */
     maxCachedSessions?: number;
-    /** TODO */
-    maxRedirections?: number;
     /** TODO */
     connect?: buildConnector.BuildOptions | buildConnector.connector;
     /** TODO */
